% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin_ctrl.R
\name{bid_ctrl}
\alias{bid_ctrl}
\title{Check Finnish Business ID (y-tunnus) validity}
\usage{
bid_ctrl(bid)
}
\arguments{
\item{bid}{a vector of 1 or more business identity numbers}
}
\description{
A function that checks whether a \code{bid} (Finnish Business ID) is valid. 
Returns \code{TRUE} or \code{FALSE}.
}
\examples{
bid_ctrl(c("0737546-2", "1572860-0")) # TRUE TRUE
bid_ctrl("0737546-1") # FALSE
}
