% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetu_diagnostic.R
\name{hetu_diagnostic}
\alias{hetu_diagnostic}
\title{Diagnostics Tool for HETU}
\usage{
hetu_diagnostic(pin, extract = NULL, subsetting = FALSE, warnings = FALSE)
}
\arguments{
\item{pin}{Finnish personal identification number as a character vector, 
or vector of identification numbers as a character vectors}

\item{extract}{Extract only selected part of the diagnostic information.
 Valid values are "\code{hetu}", "\code{is.temp}", "\code{invalid.personal.number}",
 "\code{invalid.checksum}", "\code{incorrect.checksum}", "\code{invalid.date}",
 "\code{invalid.day}", "\code{invalid.month}", "\code{invalid.length}", 
"\code{invalid.century}". If \code{NULL} (default), returns all information.}

\item{subsetting}{Print only PINs where extracted information is \code{TRUE}.}

\item{show.warnings}{If TRUE, print warnings normally. Default is FALSE, suppressing warnings.}
}
\value{
A data.frame containing PINs that have invalid parts.
}
\description{
Produce a data frame of PINs that may require closer scrutiny.
}
\examples{
diagnosis_example <- c("010101-0102", "111111-111Q", 
"010101B0101", "320101-0101", "011301-0101", 
"010101-01010", "010101-0011")
## Print all diagnoses
suppressWarnings(hetu_diagnostic(diagnosis_example))
# Extract century-related checks
suppressWarnings(hetu_diagnostic(diagnosis_example, extract = "invalid.century"))
# Extract only rows where invalid.checksum = TRUE
suppressWarnings(hetu_diagnostic(diagnosis_example, subsetting = TRUE, extract = "invalid.checksum")) 

}
