% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetu_diagnostic.R
\name{is.diagnostic}
\alias{is.diagnostic}
\title{Is an Object from Class "diagnostic"?}
\usage{
is.diagnostic(object)
}
\arguments{
\item{object}{Object to be tested}
}
\value{
TRUE or FALSE
}
\description{
Returns TRUE if the object has class "diagnostic"
}
