% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin_age.R
\name{pin_age}
\alias{pin_age}
\title{Age from ID}
\usage{
pin_age(pin, date = Sys.Date(), timespan = "years")
}
\arguments{
\item{pin}{Finnish personal identification number as a character vector, 
or vector of identification numbers as a character vectors}

\item{date}{Date at which age is calculated. If a vector is provided it must be
of the same length as the \code{pin} argument.}

\item{timespan}{Timespan to use to calculate age. The actual timespans are:
\itemize{
  \item \code{years} (Default)
  \item \code{months}
  \item \code{weeks}
  \item \code{days}
}}
}
\value{
Age as an integer vector.
}
\description{
Calculate the age in full years for a given date.
}
\examples{
ex_pin <- c("010101-0101", "111111-111C")
pin_age(ex_pin, date = "2012-01-01")

}
