% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin_ctrl.R
\name{pin_ctrl}
\alias{pin_ctrl}
\title{Finnish personal identification number validator}
\usage{
pin_ctrl(pin, allow.temp = FALSE)
}
\arguments{
\item{pin}{Finnish personal identification number as a character vector, or
vector of identification numbers as a character vectors.}

\item{allow.temp}{Allow temporary PINs (personal numbers 900-999) or only
regular PINs (personal numbers 002-899). Default is \code{FALSE}.}
}
\value{
Is the given string a valid Finnish personal identification number,
  \code{TRUE} or \code{FALSE}.
}
\description{
Validate Finnish personal identification numbers (hetu).
}
\examples{
pin_ctrl("010101-0101") # TRUE
pin_ctrl("010101-010A") # FALSE
}
\seealso{
\code{\link{hetu}} For extracting information from Finnish personal
  identification numbers.
}
\author{
Jussi Paananen
}
