% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin_date.R
\name{pin_date}
\alias{pin_date}
\alias{hetu_date}
\title{Get Birth date from PIN}
\usage{
pin_date(pin)

hetu_date(pin)
}
\arguments{
\item{pin}{Finnish personal identification number as a character vector, 
or vector of identification numbers as a character vectors}
}
\value{
Date of birth as a vector in date format.
}
\description{
Calculates the date of birth in date format.
}
\examples{
pin_date(c("010101-0101", "111111-111C"))

hetu_date(c("010101-0101", "111111-111C"))
}
