% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin_sex.R
\name{pin_sex}
\alias{pin_sex}
\title{Sex From ID}
\usage{
pin_sex(pin)
}
\arguments{
\item{pin}{Finnish personal identification number as a character vector, 
or vector of identification numbers as a character vectors}
}
\value{
Factor with label 'Male' and 'Female'.
}
\description{
Extract sex from Finnish personal identification number.
}
\examples{
pin_sex("010101-010A") 
}
\seealso{
\code{\link{hetu}} For general information extraction
}
\author{
Leo Lahti, Pyry Kantanen
}
