% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin_to_date.R
\name{pin_to_date-deprecated}
\alias{pin_to_date-deprecated}
\title{ID to Date.}
\usage{
pin_to_date(pin)
}
\arguments{
\item{pin}{Finnish ID number}
}
\value{
Date of birth as a vector in date format.
}
\description{
Calculates the date of birth in date format.
}
\seealso{
\code{\link{hetu-deprecated}}
}
\keyword{internal}
