% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetu_diagnostic.R
\name{plot.diagnostic}
\alias{plot.diagnostic}
\title{Plotting method for diagnostic class objects}
\usage{
\method{plot}{diagnostic}(x, negate.logicals = FALSE, labels = TRUE, ...)
}
\arguments{
\item{x}{a "summary.diagnostic" object}

\item{negate.logicals}{negate TRUE and FALSE logicals, default is FALSE.
Sometimes it may be beneficial to emphasize FALSE cases instead of TRUE}

\item{labels}{include column labels on y-axis, default is TRUE}

\item{...}{Arguments to be passed to methods, such as graphical parameters.
For example:
   \itemize{
     \item{\strong{type} what type of plot should be drawn.
     Default is "h" for histogram / high density vertical lines.}
     \item{\strong{lwd} line width as double Default is 1.0}
   }
   See \code{\link[base]{plot}} and \code{\link[graphics]{par}} for
   more options}
}
\description{
Creates a concise plot that visualizes TRUE and FALSE cases
in a diagnostics data frame
}
\details{
There seems to be no canonical answer on what to call this type of
plot. Some of the names that can be found online when describing a plot for
binary response value on an axis are: a one-dimensional scatterplot,
a sparkline, a rug plot, or a strip plot / strip chart.
}
