% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpin.R
\name{rbid}
\alias{rbid}
\title{Generate a vector of random Finnish Business ID's (y-tunnus)}
\usage{
rbid(n)
}
\arguments{
\item{n}{number of generated BIDs}
}
\value{
a vector of generated \code{BID}-numbers.
}
\description{
A function that generates random Finnish Business ID's, \code{bid}-numbers (Y-tunnus).
}
\examples{
x <- rbid(3)
bid_ctrl(x)
}
