% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpin.R
\name{rpin}
\alias{rpin}
\title{Generate a vector of random \code{hetu}}
\usage{
rpin(
  n,
  start_date = as.Date("1895-01-01"),
  end_date = Sys.Date(),
  p.male = 0.3,
  p.temp = 0.1
)
}
\arguments{
\item{n}{number of observations.}

\item{start_date}{Lower limit of generated \code{hetu} dates. Default is 1895-01-01.}

\item{end_date}{Upper limit of generated \code{hetu}. Default is the current date.}

\item{p.male}{Proportion of males. Default is 0.3.}

\item{p.temp}{Proportion of temporary identification numbers. Default is 0.1. Currently not implemented.}
}
\value{
a vector of generated \code{hetu}-pins.
}
\description{
A function that generates random \code{hetu}-pins. 
The generated \code{hetu} is uniformely distributed over the time period.
}
\examples{
x <- rpin(3)
hetu(x)
hetu(x, extract = "gender")
hetu(x, extract = "checksum")

}
