% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetu_control_char.R
\name{satu_control_char}
\alias{satu_control_char}
\title{Finnish Unique Identification Number Control Character Calculator}
\usage{
satu_control_char(pin, complement = FALSE)
}
\arguments{
\item{pin}{An incomplete FINUID that has 8 numbers}

\item{complement}{Should the function print only}
}
\value{
Control character, either a number 0-9 or a letter. If complete
   is TRUE, then the function returns a complete FINUID.
}
\description{
Calculate a valid control character for an incomplete 
   Finnish Unique Identification Number (FINUID, or sähköinen asiointitunnus
   SATU).
}
\details{
This method of calculating the control character was devised by
   mathematician Erkki Pale (1962) to detect input errors but also to 
   detect errors produced by early punch card machines. The long number 
   produced by writing the birth date and the personal number together are
   divided by 31 and the remainder is used to look up the control character
   from a separate table containing alphanumeric characters except letters
   G, I, O, Q and Z.
   
   The method of calculating the control character does not need century 
   character and therefore the function has an option to omit it.
}
\examples{
satu_control_char("10000001")
}
\author{
Pyry Kantanen
}
