% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin_ctrl.R
\name{satu_ctrl}
\alias{satu_ctrl}
\title{Check Validity of Finnish Unique Identification Number (SATU)}
\usage{
satu_ctrl(satu)
}
\arguments{
\item{satu}{a vector of 1 or more Unique Identification Numbers}
}
\description{
A function that checks whether a \code{satu} (Finnish Unique Identification
Number) is valid. Returns \code{TRUE} or \code{FALSE}.
}
\examples{
satu_ctrl("10000001N") # TRUE
satu_ctrl(c("10000001N", "20000001B")) # TRUE FALSE
}
