% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpin.R
\name{rpin}
\alias{rpin}
\alias{rhetu}
\title{Generate a vector of random \code{hetu}}
\usage{
rpin(
  n,
  start.date = as.Date("1895-01-01"),
  end.date = Sys.Date(),
  p.male = 0.4,
  p.temp = 0
)

rhetu(
  n,
  start.date = as.Date("1895-01-01"),
  end.date = Sys.Date(),
  p.male = 0.4,
  p.temp = 0
)
}
\arguments{
\item{n}{number of generated \code{hetu}-pins}

\item{start.date}{Lower limit of generated \code{hetu} dates,
character string in ISO 8601 standard, for example "2001-02-03".
Default is "1895-01-01".}

\item{end.date}{Upper limit of generated \code{hetu}. 
Default is current date.}

\item{p.male}{Proportion of males, between 0.0 and 1.0. Default is 0.4.}

\item{p.temp}{Proportion of temporary identification numbers, between
0.0 and 1.0. Default is 0.0.}
}
\value{
a vector of generated \code{hetu}-pins.
}
\description{
A function that generates random \code{hetu}-pins.
}
\examples{
x <- rpin(3)
hetu(x)
hetu(x, extract = "sex")
hetu(x, extract = "checksum")

x <- rhetu(3)
x
}
\author{
Pyry Kantanen, Jussi Paananen
}
